Restructure api schemas.
