+++ ansible_events_ui/api/activation.py
@@ -10,7 +10,7 @@ from fastapi import APIRouter, Depends, Response, status
 from fastapi.exceptions import HTTPException
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from ansible_events_ui import schemas
+from ansible_events_ui import schema
 from ansible_events_ui.config import Settings, get_settings
 from ansible_events_ui.db import models
 from ansible_events_ui.db.dependency import (
@@ -29,11 +29,11 @@ router = APIRouter()
 
 @router.post(
     "/api/activations/",
-    response_model=schemas.ActivationBaseRead,
+    response_model=schema.ActivationBaseRead,
     operation_id="create_activation",
 )
 async def create_activation(
-    activation: schemas.ActivationCreate,
+    activation: schema.ActivationCreate,
     db: AsyncSession = Depends(get_db_session),
 ):
     query = sa.insert(models.activations).values(
@@ -60,7 +60,7 @@ async def create_activation(
 
 @router.get(
     "/api/activation/{activation_id}",
-    response_model=schemas.ActivationRead,
+    response_model=schema.ActivationRead,
     operation_id="show_activation",
 )
 async def read_activation(
@@ -140,12 +140,12 @@ async def read_activation(
 
 @router.patch(
     "/api/activation/{activation_id}",
-    response_model=schemas.ActivationBaseRead,
+    response_model=schema.ActivationBaseRead,
     operation_id="update_activation",
 )
 async def update_activation(
     activation_id: int,
-    activation: schemas.ActivationUpdate,
+    activation: schema.ActivationUpdate,
     db: AsyncSession = Depends(get_db_session),
 ):
     stored_activation = (
@@ -208,7 +208,7 @@ async def read_output(proc, activation_instance_id, db_session_factory):
 
 @router.post("/api/activation_instance/")
 async def create_activation_instance(
-    a: schemas.ActivationInstance,
+    a: schema.ActivationInstance,
     db: AsyncSession = Depends(get_db_session),
     db_session_factory: sqlalchemy.orm.sessionmaker = Depends(
         get_db_session_factory
@@ -322,7 +322,7 @@ async def delete_activation_instance(
 
 @router.get(
     "/api/activation_instance_logs/",
-    response_model=List[schemas.ActivationLog],
+    response_model=List[schema.ActivationLog],
 )
 async def list_activation_instance_logs(
     activation_instance_id: int, db: AsyncSession = Depends(get_db_session)
