+++ ansible_events_ui/schema/extra_vars.py
@@ -1,0 +1,14 @@
+from typing import Optional
+
+from pydantic import BaseModel, StrictStr
+
+
+class Extravars(BaseModel):
+    name: StrictStr
+    extra_var: StrictStr
+    id: Optional[int]
+
+
+class ExtravarsRef(BaseModel):
+    name: StrictStr
+    id: Optional[int]
