+++ ansible_events_ui/schema/message.py
@@ -1,0 +1,34 @@
+import uuid
+from datetime import datetime
+
+from pydantic import BaseModel, StrictStr, confloat, validator
+
+
+class ProducerMessage(BaseModel):
+    name: StrictStr
+    message_id: StrictStr = ""
+    timestamp: StrictStr = ""
+    lat: confloat(gt=-90, lt=90)
+    lon: confloat(gt=-180, lt=180)
+
+    @validator("message_id", pre=True, always=True)
+    def set_id_from_name_uuid(cls, v, values):
+        if "name" in values:
+            return f"{values['name']}_{uuid.uuid4()}"
+        else:
+            raise ValueError("name not set")
+
+    @validator("timestamp", pre=True, always=True)
+    def set_datetime_utcnow(cls, v):
+        return str(datetime.utcnow())
+
+
+class ProducerResponse(BaseModel):
+    name: StrictStr
+    message_id: StrictStr
+    topic: StrictStr
+    timestamp: StrictStr = ""
+
+    @validator("timestamp", pre=True, always=True)
+    def set_datetime_utcnow(cls, v):
+        return str(datetime.utcnow())
