+++ ansible_events_ui/schema/playbook.py
@@ -1,0 +1,8 @@
+from typing import Optional
+
+from pydantic import BaseModel, StrictStr
+
+
+class PlaybookRef(BaseModel):
+    id: Optional[int]
+    name: StrictStr
